Added documentation for party handlers
I know some documentation comments might seem silly, but for consistancy
we should just document everything. And in some cases even a seemingly
silly comment can help in clearing up what's happening, from where an
action is usually initiated, etc. The other thing is the "examples",
which can be useful in finding and documenting differences in packets
for reference.