Disabled an error message
Cutscenes are finished from multiple people from time to time, in which
case the current cutscene may already be null when the second packet
gets in, which would result in the disabled error. Setting only the
sender's cutscene to null wouldn't work though, since the finish packet
is only sent if the leader's finish didn't get in first, which would
trigger the end cutscenes packet for all members.