Replaced TaggableData with Tags
To make the transition a little easier I kept the HasTag method on the
data, since there are a *ton* of references to it... Someone gotta
update that some time =< Though then the question is, do we mix HasTag
and Tags.Match calls, since we work with the data objects a lot, or do
we put a Tags object into every object? Or maybe Tags should be a
static, like Regex, and we pass two strings to it?
Close #127.