Fixed #142 - Target's skill status
The TargetAction's skill id would be set to the active skill's id, no
matter what state it was in. Officials seem to use 'None' if a skill is
currently being loaded, which was tested by hitting bears while they
were loading Counterattack. This fixes the bug, as we can safely check
for the skill id in the handlers now, and it should fix similar bugs,
that weren't discovered yet.