Fixed UnkCharWindow op
It looked like there was a second char window request added, but I
probably missed that the first one was shifted in a previous update. It
can't be a coincident that there's only one char window request sent,
and it shifted by exactly 7, the shift number of a previous update.