Added fixes for Bean Rua Guard and Observatory scripts
Removed the `close()` calls from both scripts. The `close()` call in the bean rua guard script caused all of the text within the block that `close()` was contained in to get skipped, the last line of dialogue was changed to `end().` The `close()` calls in the observatory script caused an exception after the cutscene began. Cutscenes seem to end dialogue upon beginning, so the `close()` call has been removed completely.