Added #69 - Pet body size changing with age
Made some minor changes to the body update on aging, so the body
proportions can now be defined for all creatures. The values for normal
characters can be found in the aging data now as well, which we
previously handled with a formula. With this change however, the size
changes can now be freely defined.